// src/lib/api/types.ts
export interface PolygonTickerSnapshot {
  ticker: string;
  todaysChangePerc: number;
  todaysChange: number;
  updated: number;
  day: {
    o: number; // open
    h: number; // high
    l: number; // low
    c: number; // close
    v: number; // volume
    vw: number; // volume weighted
  };
  min: {
    c: number; // close
    h: number; // high
    l: number; // low
    v: number; // volume
  };
  prevDay: {
    o: number;
    h: number;
    l: number;
    c: number;
    v: number;
  };
}

export interface PolygonAggregateBar {
  T: string; // ticker
  v: number; // volume
  vw: number; // volume weighted average price
  o: number; // open
  c: number; // close
  h: number; // high
  l: number; // low
  t: number; // timestamp
  n: number; // number of transactions
}